//홈 컨트롤러 이것저것 나눔
맵퍼 또 수정
어드민 페이지 건드릴려고 했는데 일단은 스탑

+++ STS_HighRice_NewTable/src/main/java/com/bit_fr/vo/MemberVo.java
@@ -13,7 +13,7 @@ public class MemberVo {
 	private int payback;
 	private String account_no;
 	private String bank;
-	private int balance;
+	private long balance;
 	private int grade;
 
 	public String getMember_id() {
@@ -120,11 +120,11 @@ public class MemberVo {
 		this.bank = bank;
 	}
 
-	public int getBalance() {
+	public long getBalance() {
 		return balance;
 	}
 
-	public void setBalance(int balance) {
+	public void setBalance(long balance) {
 		this.balance = balance;
 	}
 
@@ -137,7 +137,7 @@ public class MemberVo {
 	}
 
 	public MemberVo(String member_id, String pwd, String name, String tel, String jumin, String pwd_q, String pwd_a,
-			String address, String address_detail, int payback, String account_no, String bank, int balance,
+			String address, String address_detail, int payback, String account_no, String bank, long balance,
 			int grade) {
 		super();
 		this.member_id = member_id;
